llamaGen/requirements.txt
accelerate==1.6.0
aesthetic-predictor-v2-5==2024.12.18.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
appdirs==1.4.4
args==0.1.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
beautifulsoup4==4.13.4
blessed==1.21.0
boto3==1.38.27
botocore==1.38.27
braceexpand==0.1.7
branca==0.8.1
bs4==0.0.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
clean-fid==0.1.35
click==7.1.2
clint==0.5.1
cloudscraper==1.2.71
colorama==0.4.6
config==0.5.1
contourpy==1.3.2
cycler==0.12.1
datascience==0.17.6
datasets==2.21.0
debugpy==1.8.14
decorator==5.2.1
deepspeed==0.15.4
diffusers==0.33.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
folium==0.19.6
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.6.1
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-genai==1.24.0
gpustat==1.1.1
h11==0.16.0
hjson==3.1.0
hpsv2==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
icecream==2.1.7
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iniconfig==2.1.0
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
json_repair==0.47.6
kiwisolver==1.4.8
latex2sympy2_extended==1.10.1
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
math-verify==0.7.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.4
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
narwhals==1.39.1
natsort==8.4.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
nvidia-ml-py==12.575.51
open_clip_torch==2.32.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
pixi==1.0.1
pixiv-api==0.3.7
platformdirs==4.3.8
plotly==6.1.0
pluggy==1.6.0
prettytable==3.16.0
proglog==0.1.12
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyecharts==2.0.8
Pygments==2.19.1
pynvml==11.4.1
pyparsing==3.2.3
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scenedetect==0.6.6
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.28.0
setproctitle==1.3.6
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
swankit==0.2.4
swanlab==0.6.8
sympy==1.13.1
tenacity==8.5.0
timm==1.0.15
tokenizers==0.21.0
tomli==2.2.1
torch @ file:///home/bingxing2/apps/package/pytorch/2.5.1-cu121_cp310/torch-2.5.1%2Bcu121-cp310-cp310-linux_aarch64.whl#sha256=7e4602f8cad6cca70523770fcb58b567cfafacca02dd271fb67dd7e22e736703
torchaudio @ file:///home/bingxing2/apps/package/pytorch/2.5.1-cu121_cp310/torchaudio-2.5.1a0%2B1661daf-cp310-cp310-linux_aarch64.whl#sha256=bd03f3a1a372b70da63835a86c2f93c187e93477329837b9e63d7e9f24ecfe41
torchvision @ file:///home/bingxing2/apps/package/pytorch/2.5.1-cu121_cp310/torchvision-0.20.1a0%2B3ac97aa-cp310-cp310-linux_aarch64.whl#sha256=e66dea7cd5f6464299b180601ea6f61c28cbb4c9405e292bfcdfe9e6a9e42f17
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uv==0.7.9
wandb==0.19.9
wcwidth==0.2.13
webdataset==0.2.111
websockets==15.0.1
wrapt==1.17.3
xxhash==3.5.0
xyzservices==2025.4.0
yarl==1.20.0
zipp==3.21.0
